requirements.txt
flask
groq
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
requests
python-dotenv